refactor: 책임 분리
- Controller는 Service를 호출하도록 변경
- race 메소드 분리
