+++ README.md
@@ -23,3 +23,8 @@
 - [x] 입력받은 값을 이용해 실행되는 게임의 로직을 구현한다.
   - [x] 게임 내에서 자동차가 전진 또는 멈추게 한다.
 - [x] 자동차의 상태를 화면에 출력한다.
+
+피드백 반영
+- [ ] CarTest: 경계값을 테스트할 수 있도록 변경하기
+- [x] 컨트롤러가 상태를 갖는게 적절한지(Cars), 다른 컨트롤러를 갖는게 맞는지(CarController) 고민
+- [ ] CarController에서 random을 제거하고 moveAll() 테스트 가능하도록 변경하기
