+++ src/main/java/car_racing/Main.java
@@ -1,7 +1,5 @@
 package car_racing;
 
-import car_racing.domain.GameController;
-
 public class Main {
     public static void main(String[] args) {
         GameController gameController = new GameController();
